feat: course 도메인을 domain-first 구조로 정리
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
