+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/CourseDetailResponse.java
@@ -1,0 +1,11 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+import java.math.BigDecimal;
+
+public record CourseDetailResponse(
+        Long id,
+        String title,
+        BigDecimal distanceKm,
+        Integer estimatedDurationMin
+) {
+}
