+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/CourseDownloadResponse.java
@@ -1,0 +1,11 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+import java.util.List;
+
+public record CourseDownloadResponse(
+        Long courseId,
+        String name,
+        String visibility,
+        List<CourseRoutePointResponse> routePoints
+) {
+}
