+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/CourseListResponse.java
@@ -1,0 +1,10 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+import java.util.List;
+
+public record CourseListResponse(
+        List<CourseListItemResponse> items,
+        boolean hasNext,
+        String nextCursor
+) {
+}
