+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/CourseRoutePointResponse.java
@@ -1,0 +1,10 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+import java.math.BigDecimal;
+
+public record CourseRoutePointResponse(
+        Integer pointOrder,
+        BigDecimal latitude,
+        BigDecimal longitude
+) {
+}
