+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/CourseRoutePointsResponse.java
@@ -1,0 +1,9 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+import java.util.List;
+
+public record CourseRoutePointsResponse(
+        Long courseId,
+        List<CourseRoutePointResponse> points
+) {
+}
