+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/CourseShareResponse.java
@@ -1,0 +1,9 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+public record CourseShareResponse(
+        String shareType,
+        String visibility,
+        String shareUrl,
+        String shareToken
+) {
+}
