+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/CourseWriteResponse.java
@@ -1,0 +1,9 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+public record CourseWriteResponse(
+        Long courseId,
+        Long ownerUserId,
+        String visibility,
+        String title
+) {
+}
