+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/CreateCourseFromRideRecordRequest.java
@@ -1,0 +1,9 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+public record CreateCourseFromRideRecordRequest(
+        Long sourceRideRecordId,
+        String name,
+        String description,
+        String visibility
+) {
+}
