+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/FeaturedCourseResponse.java
@@ -1,0 +1,9 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+import java.util.List;
+
+public record FeaturedCourseResponse(
+        String sortingMode,
+        List<FeaturedCourseItemResponse> courses
+) {
+}
