+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/UpdateCourseRequest.java
@@ -1,0 +1,11 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+import java.util.List;
+
+public record UpdateCourseRequest(
+        String name,
+        String description,
+        String visibility,
+        List<CourseRoutePointRequest> routePoints
+) {
+}
