+++ src/main/java/com/bikeprojectminji/bikeback/course/dto/UpdateCourseVisibilityRequest.java
@@ -1,0 +1,6 @@
+package com.bikeprojectminji.bikeback.course.dto;
+
+public record UpdateCourseVisibilityRequest(
+        String visibility
+) {
+}
