+++ src/main/java/com/bikeprojectminji/bikeback/course/entity/CourseVisibility.java
@@ -1,0 +1,7 @@
+package com.bikeprojectminji.bikeback.course.entity;
+
+public enum CourseVisibility {
+    PRIVATE,
+    UNLISTED,
+    PUBLIC
+}
