--- src/main/java/com/bikeprojectminji/bikeback/dto/course/CourseDownloadResponse.java
@@ -1,11 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-import java.util.List;
-
-public record CourseDownloadResponse(
-        Long courseId,
-        String name,
-        String visibility,
-        List<CourseRoutePointResponse> routePoints
-) {
-}
