--- src/main/java/com/bikeprojectminji/bikeback/dto/course/CourseListItemResponse.java
@@ -1,11 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-import java.math.BigDecimal;
-
-public record CourseListItemResponse(
-        Long id,
-        String title,
-        BigDecimal distanceKm,
-        Integer estimatedDurationMin
-) {
-}
