--- src/main/java/com/bikeprojectminji/bikeback/dto/course/CourseListResponse.java
@@ -1,10 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-import java.util.List;
-
-public record CourseListResponse(
-        List<CourseListItemResponse> items,
-        boolean hasNext,
-        String nextCursor
-) {
-}
