--- src/main/java/com/bikeprojectminji/bikeback/dto/course/CourseRoutePointResponse.java
@@ -1,10 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-import java.math.BigDecimal;
-
-public record CourseRoutePointResponse(
-        Integer pointOrder,
-        BigDecimal latitude,
-        BigDecimal longitude
-) {
-}
