--- src/main/java/com/bikeprojectminji/bikeback/dto/course/CourseRoutePointsResponse.java
@@ -1,9 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-import java.util.List;
-
-public record CourseRoutePointsResponse(
-        Long courseId,
-        List<CourseRoutePointResponse> points
-) {
-}
