--- src/main/java/com/bikeprojectminji/bikeback/dto/course/CourseShareResponse.java
@@ -1,9 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-public record CourseShareResponse(
-        String shareType,
-        String visibility,
-        String shareUrl,
-        String shareToken
-) {
-}
