--- src/main/java/com/bikeprojectminji/bikeback/dto/course/CourseWriteResponse.java
@@ -1,9 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-public record CourseWriteResponse(
-        Long courseId,
-        Long ownerUserId,
-        String visibility,
-        String title
-) {
-}
