--- src/main/java/com/bikeprojectminji/bikeback/dto/course/CreateCourseFromRideRecordRequest.java
@@ -1,9 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-public record CreateCourseFromRideRecordRequest(
-        Long sourceRideRecordId,
-        String name,
-        String description,
-        String visibility
-) {
-}
