--- src/main/java/com/bikeprojectminji/bikeback/dto/course/FeaturedCourseResponse.java
@@ -1,9 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-import java.util.List;
-
-public record FeaturedCourseResponse(
-        String sortingMode,
-        List<FeaturedCourseItemResponse> courses
-) {
-}
