--- src/main/java/com/bikeprojectminji/bikeback/dto/course/UpdateCourseRequest.java
@@ -1,11 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-import java.util.List;
-
-public record UpdateCourseRequest(
-        String name,
-        String description,
-        String visibility,
-        List<CourseRoutePointRequest> routePoints
-) {
-}
