--- src/main/java/com/bikeprojectminji/bikeback/dto/course/UpdateCourseVisibilityRequest.java
@@ -1,6 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.course;
-
-public record UpdateCourseVisibilityRequest(
-        String visibility
-) {
-}
