--- src/main/java/com/bikeprojectminji/bikeback/entity/course/CourseRoutePointEntity.java
@@ -1,60 +1,0 @@
-package com.bikeprojectminji.bikeback.entity.course;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.math.BigDecimal;
-
-@Entity
-@Table(name = "course_route_points")
-public class CourseRoutePointEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "course_id", nullable = false)
-    private Long courseId;
-
-    @Column(name = "point_order", nullable = false)
-    private Integer pointOrder;
-
-    @Column(nullable = false, precision = 10, scale = 7)
-    private BigDecimal latitude;
-
-    @Column(nullable = false, precision = 10, scale = 7)
-    private BigDecimal longitude;
-
-    protected CourseRoutePointEntity() {
-    }
-
-    public CourseRoutePointEntity(Long courseId, Integer pointOrder, BigDecimal latitude, BigDecimal longitude) {
-        this.courseId = courseId;
-        this.pointOrder = pointOrder;
-        this.latitude = latitude;
-        this.longitude = longitude;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Long getCourseId() {
-        return courseId;
-    }
-
-    public Integer getPointOrder() {
-        return pointOrder;
-    }
-
-    public BigDecimal getLatitude() {
-        return latitude;
-    }
-
-    public BigDecimal getLongitude() {
-        return longitude;
-    }
-}
