--- src/main/java/com/bikeprojectminji/bikeback/entity/course/CourseVisibility.java
@@ -1,7 +1,0 @@
-package com.bikeprojectminji.bikeback.entity.course;
-
-public enum CourseVisibility {
-    PRIVATE,
-    UNLISTED,
-    PUBLIC
-}
